requirements.txt
flask==2.1.2
mysql-connector-python==8.0.29
requests==2.27.1